+++ config/runtime.exs
@@ -70,11 +70,17 @@ if config_env() == :prod do
   config :alchem_iiif, AlchemIiif.Ingestion.ZipProcessor,
     max_extracted_bytes:
       System.get_env("ZIP_MAX_EXTRACTED_BYTES", "2147483648")
+      |> String.to_integer(),
+    max_pages:
+      System.get_env("ZIP_MAX_PAGES", "1500")
       |> String.to_integer()
 
   config :alchem_iiif,
     max_source_upload_bytes:
       System.get_env("MAX_SOURCE_UPLOAD_BYTES", "200000000")
+      |> String.to_integer(),
+    pdf_max_pages:
+      System.get_env("PDF_MAX_PAGES", "1500")
       |> String.to_integer()
 
   # ## SSL Support